user context e componetizaação da rota timeline
