+++ src/contexts/Usercontext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+const UserContext=createContext();
+export  default UserContext;
